BPM is still a little flaky, wristboard adds 128 to the accel values before sending

+++ Processing Hub Firmware/processing_main/inc/Accel.h
@@ -10,6 +10,6 @@
 #include <stdint.h>
 uint32_t sqrt(uint32_t n);
 
-int Displacement();
+uint8_t Displacement();
 
 #endif /* FEAR_H_ */
